Fix: testing other rack config
